refactor(suite-common): Move package to wallet-core
