+++ packages/suite/src/actions/suite/__tests__/storageActions.test.ts
@@ -5,24 +5,22 @@ import * as storageActions from '../storageActions';
 import * as suiteActions from '../suiteActions';
 import * as walletSettingsActions from '@settings-actions/walletSettingsActions';
 import * as discoveryActions from '@wallet-actions/discoveryActions';
-import { disableAccountsThunk } from '@suite-common/wallet-core';
+import { disableAccountsThunk, transactionActions } from '@suite-common/wallet-core';
 import * as SUITE from '@suite-actions/constants/suiteConstants';
 
-import { accountsReducer } from '@wallet-reducers';
+import { accountsReducer, transactionsReducer } from '@wallet-reducers';
 import walletSettingsReducer from '@wallet-reducers/settingsReducer';
 import suiteReducer from '@suite-reducers/suiteReducer';
 import deviceReducer from '@suite-reducers/deviceReducer';
 import discoveryReducer from '@wallet-reducers/discoveryReducer';
 import sendFormReducer from '@wallet-reducers/sendFormReducer';
 import graphReducer from '@wallet-reducers/graphReducer';
-import { transactionsReducer } from '@wallet-reducers';
 import fiatRatesReducer from '@wallet-reducers/fiatRatesReducer';
 import storageMiddleware from '@wallet-middlewares/storageMiddleware';
 import { getAccountTransactions, getAccountIdentifier } from '@suite-common/wallet-utils';
 import { AppState } from '@suite-types';
 import { SETTINGS } from '@suite/config/suite';
 import { preloadStore } from '@suite-support/preloadStore';
-import { transactionActions } from '@suite-common/wallet-transactions';
 
 const { getSuiteDevice, getWalletAccount, getWalletTransaction } = global.JestMocks;
