+++ packages/suite/src/middlewares/wallet/fiatRatesMiddleware.ts
@@ -1,10 +1,9 @@
 import { MiddlewareAPI } from 'redux';
 import { BLOCKCHAIN } from '@wallet-actions/constants';
 import { WALLET_SETTINGS } from '@settings-actions/constants';
-import { accountsActions } from '@suite-common/wallet-core';
+import { accountsActions, transactionActions } from '@suite-common/wallet-core';
 import * as fiatRatesActions from '@wallet-actions/fiatRatesActions';
 import { AppState, Action, Dispatch } from '@suite-types';
-import { transactionActions } from '@suite-common/wallet-transactions';
 
 const fiatRatesMiddleware =
     (api: MiddlewareAPI<Dispatch, AppState>) =>
@@ -14,49 +13,49 @@ const fiatRatesMiddleware =
         // pass action
         next(action);
 
-      if (accountsActions.updateAccount.match(action)) {
-        // fetch rates for new tokens added on account update
-        const account = action.payload;
-        const prevAccount = prevState.wallet.accounts.find(
-          a => a.descriptor === account.descriptor && a.symbol === account.symbol,
-        );
+        if (accountsActions.updateAccount.match(action)) {
+            // fetch rates for new tokens added on account update
+            const account = action.payload;
+            const prevAccount = prevState.wallet.accounts.find(
+                a => a.descriptor === account.descriptor && a.symbol === account.symbol,
+            );
 
-        if (account.tokens) {
-          const difference = account.tokens.filter(
-            token =>
-              !prevAccount?.tokens?.find(prevToken => prevToken.symbol === token.symbol),
-          );
+            if (account.tokens) {
+                const difference = account.tokens.filter(
+                    token =>
+                        !prevAccount?.tokens?.find(prevToken => prevToken.symbol === token.symbol),
+                );
 
-          difference.forEach(token => {
-            if (token.symbol) {
-              api.dispatch(
-                fiatRatesActions.updateCurrentRates({
-                  symbol: token.symbol,
-                  mainNetworkSymbol: account.symbol,
-                  tokenAddress: token.address,
-                }),
-              );
+                difference.forEach(token => {
+                    if (token.symbol) {
+                        api.dispatch(
+                            fiatRatesActions.updateCurrentRates({
+                                symbol: token.symbol,
+                                mainNetworkSymbol: account.symbol,
+                                tokenAddress: token.address,
+                            }),
+                        );
+                    }
+                });
             }
-          });
         }
-      }
 
-      if (accountsActions.createAccount.match(action)) {
-        // fetch current rates for account's tokens
-        const account = action.payload;
-        account.tokens?.forEach(token => {
-          if (!token.symbol) {
-            return;
-          }
-          api.dispatch(
-            fiatRatesActions.updateCurrentRates({
-              symbol: token.symbol,
-              mainNetworkSymbol: account.symbol,
-              tokenAddress: token.address,
-            }),
-          );
-        });
-      }
+        if (accountsActions.createAccount.match(action)) {
+            // fetch current rates for account's tokens
+            const account = action.payload;
+            account.tokens?.forEach(token => {
+                if (!token.symbol) {
+                    return;
+                }
+                api.dispatch(
+                    fiatRatesActions.updateCurrentRates({
+                        symbol: token.symbol,
+                        mainNetworkSymbol: account.symbol,
+                        tokenAddress: token.address,
+                    }),
+                );
+            });
+        }
 
         switch (action.type) {
             case BLOCKCHAIN.CONNECTED:
