+++ packages/suite/src/types/suite/index.ts
@@ -25,7 +25,7 @@ import type { MessageSystemAction } from '@suite-actions/messageSystemActions';
 import type { GuideAction } from '@suite-actions/guideActions';
 import type { EnvironmentType } from '@suite-common/suite-types';
 import type { Route } from '@suite-constants/routes';
-import { transactionActions } from '@suite-common/wallet-transactions';
+import { transactionActions } from '@suite-common/wallet-core';
 
 // reexport
 export type { ExtendedMessageDescriptor } from '@suite-components/Translation/components/BaseTranslation';
