+++ packages/suite/tsconfig.json
@@ -48,9 +48,6 @@
             "path": "../../suite-common/wallet-core"
         },
         {
-            "path": "../../suite-common/wallet-transactions"
-        },
-        {
             "path": "../../suite-common/wallet-types"
         },
         {
