+++ suite-common/wallet-core/src/index.ts
@@ -1,3 +1,6 @@
 export * from './accounts/accountsActions';
 export * from './accounts/accountsReducer';
 export * from './accounts/accountsThunks';
+export * from './transactions/transactionActions';
+export * from './transactions/transactionReducer';
+export * from './transactions/transactionThunks';
