+++ suite-common/wallet-core/src/transactions/transactionConstants.ts
