--- suite-common/wallet-transactions/package.json
@@ -1,22 +1,0 @@
-{
-    "name": "@suite-common/wallet-transactions",
-    "version": "1.0.0",
-    "private": true,
-    "license": "See LICENSE.md in repo root",
-    "sideEffects": false,
-    "main": "src/index",
-    "scripts": {
-        "lint": "eslint '**/*.{ts,tsx,js}'",
-        "type-check": "tsc --build"
-    },
-    "dependencies": {
-        "@reduxjs/toolkit": "^1.8.3",
-        "@suite-common/redux-utils": "*",
-        "@suite-common/suite-config": "*",
-        "@suite-common/suite-utils": "*",
-        "@suite-common/wallet-types": "*",
-        "@suite-common/wallet-utils": "*",
-        "@trezor/connect": "9.0.0-beta.6",
-        "file-saver": "^2.0.5"
-    }
-}
