--- suite-common/wallet-transactions/src/index.ts
@@ -1,3 +1,0 @@
-export * from './transactionReducer';
-export * from './transactionActions';
-export * from './transactionThunks';
