--- suite-common/wallet-transactions/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": { "outDir": "libDev" },
-    "references": [
-        { "path": "../redux-utils" },
-        { "path": "../suite-config" },
-        { "path": "../suite-utils" },
-        { "path": "../wallet-types" },
-        { "path": "../wallet-utils" }
-    ]
-}
